requirements.txt
discord.py==2.3.2
MetaTrader5>=5.0.39
pandas>=2.0.0
numpy>=1.25.0
matplotlib>=3.7.0
mplfinance>=0.12.9a0
cryptography>=41.0.0
python-dotenv>=1.0.0
asyncio
aiohttp>=3.8.0
audioop-lts>=0.2.2
plotly>=5.0.0
